server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
sqlalchemy==2.0.25
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiofiles==23.2.1
google-generativeai==0.3.2
deep-translator==1.11.4
pydantic==2.5.3
httpx==0.26.0
